+++ alertmanager/remote.go
@@ -6,23 +6,10 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/url"
-	"path"
 	"time"
 
 	log "github.com/Sirupsen/logrus"
 )
-
-// joinURL can be used to join a base url (http(s)://domain.com) and a path (/my/path)
-// it will return a joined string or an error (if you supply invalid url)
-func joinURL(base string, sub string) (string, error) {
-	u, err := url.Parse(base)
-	if err != nil {
-		return "", err
-	}
-	u.Path = path.Join(u.Path, sub)
-	return u.String(), nil
-}
 
 // getJSONFromURL is a helper function that takesan URL, request timeout
 // and target structure, it will make a HTTP request and decode JSON response
